fixes on regexps and fews addons
File renamed to mapper_erlang.ts (typescript), syntax for .erl and .hrl files, icon changes and few minor fixes on functions and defines regexps
Added "-include()"   statements syntax
Added "-behaviour()" statements syntax
FIX: Function arguments in many regexps: functions defines, includes, char '|' is reserved in CodeMap but a possible char in between Erlang function params
